backend/pyproject.toml
[tool.poetry]
name = "opensplit-backend"
version = "0.1.0"
description = "OpenSplit - Bill Splitting App Backend"
authors = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.32.0", extras = ["standard"] }
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.0"
pydantic = { version = "^2.8.0", extras = ["email"] }
pydantic-settings = "^2.4.0"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-multipart = "^0.0.12"
httpx = "^0.27.0"
celery = { version = "^5.4.0", extras = ["redis"] }
redis = "^5.1.0"
supabase = "^2.7.0"
llama-index = "^0.11.0"
llama-index-llms-openai = "^0.2.0"
llama-index-embeddings-openai = "^0.2.0"
openai = "^1.47.0"
pillow = "^10.4.0"
pytesseract = "^0.3.13"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
ruff = "^0.6.0"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
